refactor: split REST schema helper template
