+++ .sampo/changesets/1016-split-rest-schema-helper-template.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split the generated REST schema helper PHP template emitter from REST resource PHP template generation while preserving compatibility exports.
